Create Display and Get methods
I'm looking to make the act of getting and displaying repos and forks
repeatable, since I'd like to provide the ability to navigate between
the forks of a given repo. In the interest of that, I've refactored each
Get and Display of these into reusable functions.

Created a display method with some key info on the repo selected

Separated "GetRepo" into its own method

Separated Fork display for loop into its own method (DisplayForkTable),
and organized it into a table rather than a list of descriptive
sentences