Add Navigation between Forked Repos
I added a few prompts to allow a user to select a fork that they want to
make the focus. This also builds a hash table that builds a relationship
with the selected fork and its "Parent" repository. That way, if a user
wants to trace back up to parent repositories, they can as long as they
have been there before in this session. I would think the parent repo is
accessible somewhere I haven't seen in the API, so a future enhancement
may render this unnecessary.

TODO:

Add option to open the focused repo in a web browser

Write in some validation for user input - right now a user can break
this very easily

Allow the Full Path of the first repository to be passed into the script
as a method

Update Documentation in README.md